Rebalancing and tweaked AI
Build 0.4.1

Allowed AI player to be in player 1 slot.

Changed hardAI to ensure that when winCondition == 0, the AI will only buy an industrial district if it can afford the influence hit +10 or it has fewer than 2 industrial districts.

Changed hardAI to not Lobby if it thinks it can win soon when winCondition == 0.

Increased residential influence gain to 10.